requirements.txt
requests>=2.28.0
aiohttp>=3.8.0
networkx>=3.0
pydantic>=2.0.0
ipfshttpclient==0.8.0a2
eth-hash[pycryptodome]>=0.5.1
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# 内容审查引擎
Pillow>=10.0.0
pytesseract>=0.3.10
jieba>=0.42.1
filetype>=1.2.0  # 用于通过二进制特征准确判断文件类型